refactor: remigrate createOrUpgradeAgentConfiguration to front/lib/api/assistant/configuration.ts
